refactor: make `wait_for` an own module
closes #13
